+++ src/apple_compressor_manager/compression/__init__.py
@@ -1,6 +1,6 @@
 # src/apple_compressor_manager/single_compression/__init__.py
 
-from .single_file_compressor import compress
+from .single_profile import compress
 
 __all__ = [
     "compress",
