+++ src/apple_compressor_manager/compression/single_profile.py
